DEBBase.jl: A basis for Dynamic Energy Budget modelling in Julia

DEBBase includes a number of submodules.

The core of the package is formed by DEBBase.DEBODE and DEBBase.ABC.

DEBODE defines a baseline DEB-TKTD model and provides the core functionality to run simulations. ABC contains code for parameter estimation using Sequential Monte Carlo Approximate Bayesian Computation.

DEBODE: DEB-TKTD model as system of ordinary differential equations

DEBBase.DEBODE.DEBParamCollection — Type

A DEBParamCollection contains global parameters glb and spc parameters spc (including TKTD-parameters).
Initialize the default parameter collection with DEBParamCollection().

DEBBase.DEBODE.GlobalParams — Type

GlobalParams contain the global parameters (simulated timespan t_max, nutrient influx rate Xdot_in, etc.)

DEBBase.DEBODE.ODEAgentParams — Type

ODEAgentParams(spc::AbstractParams)

ODEAgentParams are subject to agent variability. This is in contrast to SpeciesParams, which define parameters on the species-level.

DEBBase.DEBODE.SpeciesParams — Type

SpeciesParams contain population means of DEB and TKTD parameters. Default values are roughly reproduce life-history of Daphnia minda (model currency mass carbon in μgC) exposued to Azoxystrobin (mg/L).
DEBBase.jl uses a hierarchical modelling approach where the SpeciesParams are parameters which are common across all agents of a species, and ODEAgentParams contain parameters which are specific for an individual.
Variability is given by the zoom factor Z::Distribution, which is always applied to the surface-area specific ingestion rates and can optionally propagate to parameters indicated in propagate_zoom::NTuple.
Z is Dirac(1) by default, i.e. there is no agent variability in the default parameters.

DEBBase.DEBODE.DEBODE_DA! — Method

DEBODE_DA!()

Base model assuming Damage Addition (DA) for mixture toxicity.

DEBBase.DEBODE.DEBODE_IA! — Method

DEBODE_IA!(du, u, p, t)

Definition of base model as a system of ordinary differential equations. This model definition is suitable for simulating the life-history of a single organism in conjecture with OrdinaryDiffEq/DifferentialEquations.jl.

DEBBase.DEBODE.Hbj — Method

Mautrity-driven metabolic acceleration from birth to maturity threshold H_j (metamorphosis). We assume that some baseline parameter p has value p_b at birth and p_j at metamorphosis. Between birth and metamorphosis, the current value of p is the maturity-weighted mean of p_b and p_j.

DEBBase.DEBODE.Sdot_negative — Method

Negative somatic growth

DEBBase.DEBODE.Sdot_positive — Method

Positive somatic growth

DEBBase.DEBODE.calc_SL_max — Method

calc_SL_max(spc::AbstractSpeciesParams)::Float64

Calculate maximum structural length $slmax$, given in dimension $m^(1/3)$.

DEBBase.DEBODE.calc_S_max — Method

calc_S_max(spc::AbstractParams)::Float64

Calculate maximum structural mass smax [m]

DEBBase.DEBODE.clipneg — Method

Clip negative values.

DEBBase.DEBODE.dA! — Method

Assimilation flux

DEBBase.DEBODE.dC_W! — Method

dC_W!(
    du::ComponentVector,
    u::ComponentVector,
    p::AbstractParamCollection,
    t::Real
    )::Nothing

Change in external concentrations. Currently simply returns zeros because time-variable exposure is not yet implemented.

TODO: extend to account for time-variable exposure.

DEBBase.DEBODE.dD! — Method

TK for spc-TKTD model, including effect of surface area to volume ratio and dilution by growth. If D and is given as a Vector, TK is only stressor-specific but not PMoA-specific.

DEBBase.DEBODE.dH! — Method

Maturation flux.

Maturity is dissipated energy and can therefore not be burned to cover maintenance costs.

Currently there are no consequences for an organism not covering maturity maintenance. If this turns out to be an issue, we might consider to add a damage pool $D_H$, where the amount of maturity maintenance that could not be covered is accumulated. This might then lead to a fitness penalty depending on $D_H$, for example in the form of additional mortality or embryonic hazard.

Such rules are however likely species-specific and should be evaluated in the light of a more precise problem definition.

DEBBase.DEBODE.dH_b! — Method

Update the current estimate of Hb. The current estimate of maturity at birth is equal to current maturity for embryos, and will be fixed to the current value upon completion of embryonic development. This way, we obtain Hb as an implied trait and can use it later (for example in abj()).

DEBBase.DEBODE.dI! — Method

dI!(
    du::ComponentArray,
    u::ComponentArray,
    p::AbstractParamCollection,
    t::Real
    )::Nothing

Calculate ingestion rate. Embryos (X_emb <= 0) take up resources from the vitellus X_emb. Juveniles and adults (X_emb > 0) feed on the external resource X_p.

The DEBBase model assumes that all state variables are given as mass (or a linearly proportional quantity). As this includes structure, the ingestion rate scales with the structural mass $S^{2/3}$`.

DEBBase.DEBODE.dJ! — Method

Maturity maintenance flux

DEBBase.DEBODE.dM! — Method

Somatic maintenance flux

DEBBase.DEBODE.dQ! — Method

dQ!(
    du::ComponentArray,
    u::ComponentArray,
    p::AbstractParamCollection,
    t::Real
    )::Nothing

Calculation of the total dissipation flux, equal to the sum of maintenance costs and overheads paid for assimilation, mobilization, maturation, growth and reproduction.

DEBBase.DEBODE.dR! — Method

Reproduction flux.

DEBBase.DEBODE.dS! — Method

Somatic growth, including application of shrinking equation.

DEBBase.DEBODE.dS_max_hist! — Method

dS_max_hist!(
    du::ComponentArray,
    u::ComponentArray,
    p::AbstractParamCollection,
    t::Real
    )::Nothing

Reference structure S_max_hist, which is used as a measure of energetic state.

This is the amount of structure an individual of the given age has under ideal conditions, i.e. f = 1 and y_z = 1, with the exception of y_G. Values of y_G != 1 are included in the calculation of S_max_hist, so that a slowing down of growth

DEBBase.DEBODE.dX_p! — Method

Change in environmental resource abundance, simulating a chemostat.

DEBBase.DEBODE.functional_response — Method

functional_response(
    du::ComponentArray,
    u::ComponentArray,
    p::AbstractParamCollection,
    t::Real
    )::Float64

Compute scaled functional response f_X, assuming a Holling Type II relationship.

DEBBase.DEBODE.individual_variability! — Method

individual_variability!(p::Ref{AbstractParams})

Induce agent variability in spc parameters via zoom factor Z. Z is sampled from the corresponding distribution given in p and assumed to represent a ratio between maximum structurel masses (not lengths), so that the surface area-specific ingestion rate Idot_max_rel scales with Z^(1/3) and parameters which represent masses or energy pools scales with Z.

DEBBase.DEBODE.initialize_statevars — Method

initialize_statevars(p::AbstractParamCollection, pindt::ComponentVector{Float64})::ComponentArray

For initialization of ODE simulator, initialize the component vector of state variables, u, based on common oaraeter collection theta.

DEBBase.DEBODE.k_J! — Method

k_J!(spc::AbstractSpeciesParams)::Nothing

Set the maturity maintenance rate constant, assuming that the cumulative investment into maturity maintenance equals the cumulative investment into somatic maintenance.

DEBBase.DEBODE.sig — Method

function sig( x::Float64, xthr::Float64, yleft::Float64, y_right::Float64; beta::Float64 = 1e16 )::Float6

Sigmoid switch function. This can be useful to replace simple if-statements with a continuous function.

y_left and y_right are the function values left and right of the threshold x_thr.

DEBBase.DEBODE.simulator — Method

simulator( p::AbstractParamCollection; model = DEBODE!, alg = Tsit5(), saveat = 1, reltol = 1e-6, AgentParamType::DataType = ODEAgentParams, kwargs...)::DataFrame

Run an ODE-based model.

args:

theta::AbstractParamCollection: A parameter collection with defined global parameters (<: AbstractGlobalParams) and species parameters (<: AbstractSpeciesParams).

kwargs:

model = DEBODE!: Definition of the derivatives. A function form du!(du, u, p, t)::Nothing. See definition of DEBODE! in derivatives.jl for an example, or see docs and examples of DifferentialEquations.jl for more details.

alg = Tsit5(): Algorithm to be used by solve function

saveat = 1: When or how often to save ODE solutions to output

reltol = 1e-6: Relative tolerance of ODE solutions

AgentParamType::DataType = ODEAgentParams: The data type that stores those parameters that are affected by individual variability. There has to be a corresponding constructor so that theta.agn = AgentParamType(theta.spc) works.

kwargs...: Additional keyword argument are passed on to OrdinaryDiffEq.solve

Example:

sim = simulator(DEBParamCollection())

DEBBase.DEBODE.y_z_DamageAddition! — Method

Response to chemical stressors, assuming damage addition for mixtures.
The PMoA-specific damage values for each stressor are added up,

$D_j = \sum_{z=1}^{n}{D_{j,z}}$

, and the response is caluclated assuming identical DRC parameters for all stressors. Currently, there are no additional weight factors implemented (assuming that the model is fitted to single-substance data only).

DEBBase.DEBODE.y_z_IA! — Method

Response to chemical stressors, assuming independent action for mixtures.

DEBBase.DEBODE.@replicates — Macro

@replicates(simcall::Expr, nreps::Int64)

Perform replicated runs of simcall, where simcall is a call to a simulator function.

Example:

spc = SpeciesParams(Z = Truncated(Normal(1, 0.1), 0, Inf)) # initialize default parameters with variable zoom factor
sim = @replicates DEBBase.simulator(DEBParamCollection(spc = spc))) 10 # execute replicated runs to simulator

In this case, sim will contain the output of 10 replicated simulations. For each replicate, the zoom factor is sampled from a truncated Normal distribution. sim contains an additional column replicate.

ABC: Parameter inference using Sequential Monte Carlo Approximate Bayesian Computation

DEBBase.ABC — Module

Submodule for parameter estimation using approximate bayesian computation.

Base.Iterators.zip — Method

zip(priors::Priors)

Simultaneously iterate over parameter names and prior distributions.

Base.get — Method

get(priors::Priors, param::Symbol)

Get prior distribution from priors object.

Examples:

prior_kM = get(priors, :k_M)

Base.rand — Method

rand(defaultparams::DataType, priors::Priors)

Sample from priors and assign to new instance of default parameter structure.

DEBBase.ABC.SMC — Method

SMC(
    priors::Union{Priors,Tuple{Priors,DataFrame}},
    defaultparams,
    simulator,
    distance,
    data::Any;
    n_pop::Int64 = 1000,
    q_eps::Float64 = 0.2,
    k_max::Int64 = 3,
    convergence_eps::Float64 = 0.1,
    paralellization_type = threaded,
    savetag = "smc",
    saveto::Union{String,False} = ""
    )

kwargs

convergence_eps::Float64 : we consider SMC converged if the relative difference between two successive distance thresholds ϵ is less than convergence_eps

savedata::Bool : whether to include the value of positional argument data in the output metadata. false by default

saveto::Bool : where to store the SMCResult object on disc. tempname() by default.

DEBBase.ABC.assign! — Method

assign!(particle::AbstractParams, paramname::Symbol, value::Float64; assignment_instructions::Nothing)

Assign a sample to a param struct. This accounts for the possibility that some parameters might be stored in vectors (e.g. TKTD parameters where each vector element corresponds to a chemical). If kwarg assignment_instructions is Nothing, we can ignore it.

DEBBase.ABC.assign! — Method

assign!(particle::AbstractParams, paramnames::Union{Vector{Symbol},Vector{String}}, pvals::Union{AbstractVector,DataFrameRow})

Assign new values to a parameter structure, given a list of names and values.

DEBBase.ABC.bestfit — Method

bestfit(defparams::AbstractParams, accepted::AbstractDataFrame)

Get the best fit from accepted (particle with minimum distance) and assign to a copy of defparams.

DEBBase.ABC.calcscale — Method

Calculate quantile-based parameter scale for scalar parameters.

DEBBase.ABC.calculatescales — Method

Calculate perturbation kernel scales from a Vector of accepted particles.

DEBBase.ABC.calculateweights_threaded — Method

Calculate SMC sampling weights.

DEBBase.ABC.deflognorm — Method

deflognorm(modus, sigma)

Define a truncated normal distribution based on the mode and σ. The parameter μ is then calculated as $\mu = ln(mode) + \sigma^2$.

DEBBase.ABC.deftruncnorm — Method

deftruncnorm(mu, CV; l = 0, u = Inf)

Define a truncated normal distribution based on the mean and CV.

DEBBase.ABC.extractfields — Method

Extract fields from a parameter structure.

DEBBase.ABC.gaussiankernel — Method

gaussiankernel(old_value::Float64, scale::Float64, min::Float64, max::Float64)

Define a Gaussian perturbation kernel.

DEBBase.ABC.getfieldname — Method

getfieldname(paramname::Symbol)::Symbol

Extract a valid param struct fieldname from a parameter name by dropping the suffix.

Examples

getfieldname(:k_D_A_1) # returns :k_D_A

DEBBase.ABC.getparindex — Method

getparindex(paramname::Symbol)::Int64

Infer index from a parameter name.

Examples

getparindex(:k_D_G_1) # returns 1, because parameter name ends on _1
getparindex(:k_D_G) # will fail, because no index indicated in suffix

DEBBase.ABC.iman_conover — Method

iman_conover(ar, cor_mat)

Induce rank correlations to independent samples, using Iman-Conover algorithm. This function was adopted from the MCHammer package. –- **MCHammer was released with the following License statement: **

Copyright 2019-2021, Technology Partnerz Ltd. and Eric Torkia

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

DEBBase.ABC.initialize_threaded — Method

initialize( n_pop::Int64, defaultparams::Union{AbstractParams,AbstractParamCollection}, priors::Priors, simulator, distance, data::AbstractDataFrame )

Initialization of a population if the priors are purely parameteric priors.

DEBBase.ABC.initialize_threaded — Method

initialize(
    n_pop::Int64, 
    defaultparams::AbstractParams, 
    priors::Tuple{Priors,DataFrame},
    simulator,
    distance,
    data::AbstractDataFrame
)

Initialization of the SMC population if the given priors are a combination of parameteric priors and previously accepted particles.
Note that parameteric prior distributions should always be given for all parameters. For the parameters which also appear in the dataframe of accepted values

DEBBase.ABC.isvecparam — Method

isvecparam(paramname::Symbol, particle::AbstractParams)::Bool

Check whether parameter represents a Vector element. We assume that the indices of such parameters are indicated by a suffix following the form param_i. E.g. k_D_G_1 for the first value in the vector of k_D_G-values. isvecparam checks whether such a field occurs in AbstractParams and whether it is a subtype of AbstractVector.

DEBBase.ABC.posterior_sample! — Method

Take posterior sample from a data frame of accepted values and assign to parameter structure. ID cols are at the end of the data frame and will be omitted from the sample, according to num_idcols.

DEBBase.ABC.posterior_sample — Method

posterior_sample(accepted::DataFrame; reserved_colnames::Vector{String} = ["distance", "weight", "model", "chain"])

Take posterior sample from a data frame of accepted values.

DEBBase.ABC.ppc — Method

ppc(fit::SMCResult; n_samples = 1000, kwargs...)

Run a posterior predictive check, given SMCResult object fit.

DEBBase.ABC.ppc — Method

ppc(defaultparams::AbstractParams, simulator, accepted::AbstractDataFrame, priors::Priors; n_samples = 1000)

Compute posterior predictions for posterior predictive check.

defaultparam::AbstractParams: default parameters used in simulator

simulator: simulator function with signature simulator(defaultparams, priors.params, sample), where

accepted::AbstractDataFrame: accepted particles as returned by SMC

This function assumes that simulator returns a DataFrame.

kwargs

n_samples = 1000: number of samples from accepted_particles to evaluate

DEBBase.ABC.pvec — Method

Given parameter param, convert vector of parameter structures containg param into vector of values of param.

DEBBase.ABC.rand! — Method

Sample from priors and assign to existing instance of parameter structure.

DEBBase.ABC.resample — Method

resample(
    accepted_particles::Vector{Vector{Float64}}, 
    accepted_distances::Vector{Float64},
    accepted_weights::Vector{Float64},
    priors::Priors,
    n_pop::Int64,
    parnames::Vector{Symbol},
    num_params::Int64
    )

Resample from previously accepted particles, with account for SMC sampling weights.

DEBBase.ABC.summarize_accepted — Method

summarize_accepted(
    accepted::D; 
    ffun = fround
    ) where D <: AbstractDataFrame

Compute summary statistics from a data frame of accepted particles. Values are formatted using function $ffun$.

Use $ffun = x -> x$ to suppress any kind of formatting. By default, values are rounded to 2 significant digits and formatted as strings using fround.

DEBBase.ABC.wkde — Method

function wkde(accepted::DataFrame, param::Symbol)

Compute a weighted kernel density estimate for parameter param in accepted particles accepted.

DEBBase.ABC.wkde — Method

function wkde(X::Matrix{R}, ω = Vector{R}) where R <: Real

Compute a weighted kernel density estimate on Matrix X.

DEBBase.ABC.wkde — Method

wkde(X::Vector{R}, ω = Vector{R}) where R <: Real

Compute a weighted Kernel Density Estimate of values in X with weights ω.

DoseResponse: A collection of Dose-Response functions

DEBBase.DoseResponse.CRS4 — Method

4-parameter CRS model. Lower limit and response in the control are fixed to 0 and 1, respectively.

DEBBase.DoseResponse.CRS4U — Method

4-parameter CRS model transformed to u-shape, where the response in the control is 1, the lower limit is 0 and the maximum is 2.0.

DEBBase.DoseResponse.CRS5US — Method

Re-scaled U-shaped CRS model with parameter C fixed to 0.
Parameters are
- alpha: rate of hormetic increase

b: slope of the inclining part of the curve

d: maximum stress

e: inflection point of the inclining part of the curve

f: hormesis parameter

DEBBase.DoseResponse.CRS5US — Method

DEBBase.DoseResponse.CRS6 — Method

Cedergreend-Ritz-Streibig model.
Parameters are
- p[1] = α = rate of hormetic increase

p[2] = b = quasi-slope

p[3] = c = lower limit

p[4] = d = response in the control

p[5] = e = inflection point

p[6] = f = hormesis parameter

DEBBase.DoseResponse.CRS6U — Method

6-parameter U-shaped CRS model.

DEBBase.DoseResponse.CRS6US — Method

Re-scaled 6-parameter U-shaped CRS model.

DEBBase.DoseResponse.LL2 — Method

Two-parameter log-logistic function. The expression y = 1 / (1 + (x/p[1])^p[2]) is extended to y = (1 / (1 + Complex(x/p[1])^p[2])).re. This way we deal with domain errors which might occur if x or p[1] temporarily takes a negative value. Negative values should theoretically be impossible, but very small values of x (<= 1e-20) might occur during ODE solving. In this case, the returned real part of the expression evaluates to 1, which is in turn the expected behaviour.

DEBBase.DoseResponse.LL2M — Method

Two-parameter log-logistic function transformed to increasing function for application to PMoA maintenance costs.

DEBBase.DoseResponse.LL2M — Method

DEBBase.DoseResponse.LL2h — Method

Cumulative hazard function of the log-logistic distribution. Mainly used for application in GUTS.

DEBBase.DoseResponse.LL2hinv — Method

Inverse of the cumulative hazard two-parameter log-logistic function.

DEBBase.DoseResponse.LL2inv — Method

Inverse of the two-parameter log-logistic function.

DEBBase.DoseResponse.LLAS3 — Method

Asymmetric log-logistic function with additional slope parameter. Arguments: - p1 = EC50 - p2 = beta - p3 = beta_2

DEBBase.DoseResponse.LLBP5 — Method

Bi-phasic log-logistic function, where each phase is described by a two-parameter log-logistic function. Requires additional "breakpoint" parameter. Arguments: - p1 = EC501 - p2 = beta1 - p3 = EC502 - p4 = beta2 - p5 = breakpoint = relative response at which second phase starts

DEBBase.DoseResponse.NEC2neg — Method

Decreasing NEC model.

DEBBase.DoseResponse.NEC2pos — Method

Increasing NEC model.

DEBBase.DoseResponse.WB2 — Method

Two-parameter Weibull function.

DEBBase.DoseResponse.contmax — Method

Continuous maximum function, using a sigmoid function to determine the maximum.

Figures: Recipes and utilities for plotting model output

DEBBase.Figures — Module

Recipes and convenience functions for plotting model output.

DEBBase.Figures.gridxattr — Method

gridxattr(attr, nullattr, numrows::Int64, numcols::Int64)::Matrix{String}

Analogous to gridyattr.

DEBBase.Figures.gridyattr — Method

Assign some attribute value attr to the left-most column of a multi-plot and another attribute value nullattr to all other subplots.

Example:

using Plots, Plots.Measures
# sets the left left column margin to 2.5mm and the other ones to 0mm
plot(plot(), plot(), plot(), leftmargin = gridyattr(2.5mm, 0mm, 1, 3), layout = (1,3)) 

DEBBase.Figures.gridylabel — Method

Create shared y-labels for plot with grid layout, placing labels only on the left-most subplots.

DEBBase.Figures.saveplt — Method

Save plot plt with name, assuming known global TAG and plotsdir.

ParamStructs: Infrastructure to organize parameter structures

By default, DEBBase uses mutable structs to store parameters. <br>

Utils: Other utility functions

DEBBase.Utils.diffvec — Method

diffvec(x::AbstractVector)

Calculate difference along a vector, inserting NaN as first element (analogous to Pandas' diff method).

DEBBase.Utils.drop_na! — Method

drop_na!(df::DataFrame)::Nothing

Mutating version of drop_na

DEBBase.Utils.drop_na — Method

drop_na(df::AbstractDataFrame; verbose=false)::DataFrame

Drop all rows with missing values from data frame.

DEBBase.Utils.extract_colnames — Method

extract_colnames(u::ComponentVector)

Extract the column names of an output dataframe from an ODE solution object.

DEBBase.Utils.extract_colnames — Method

extract_colnames(sol::O)::Vector{Symbol} where {O <:ODESolution}

Extract the column names of an output dataframe from an ODE solution object.

DEBBase.Utils.fround — Method

fround(x; sigdigits=2)

Formatted rounding to significant digits (omitting decimal point when appropriate). Returns rounded number as string.

DEBBase.Utils.geomrange — Method

geomrange(a::Real, b::Real; length=50)

Geometric series created from two extreme values.

DEBBase.Utils.geomrange — Method

geomrange(v::Vector{Float64}; length=50)

Geometric series created from range of values within a vector.

DEBBase.Utils.get_treatment_names — Method

get_treatment_names(exposure::Vector{Vector{Float64}}, stressor_names::Array{Symbol,1})

Infer treatment types (categorical), levels (ordinal) and names (type + level) from Array of exposure concentrations.

DEBBase.Utils.getstat — Method

getstat(data::Vector{DataFrame}, statistic::String)

Retrieve the mean of a statistic from a data frame with columns "statistic" (listing different kinds of summary statistics) and "value".

DEBBase.Utils.idcol! — Method

idcol!(sim::Any, col::Symbol, val)

Add identifier column to a results object sim, assuming that sim is some collection of DataFrames (e.g. Tuple, Vector).

DEBBase.Utils.idcol! — Method

idcol!(sim::AbstractDataFrame, col::Symbol, val)

Add identifier column with name col and value val to results DataFrame sim

DEBBase.Utils.into! — Method

into!(sink::AbstractParams, source::AbstractParams)::Nothing

Put values from source into sink.

source::AbstractParams: A param struct from which we want to extract all parameters.

sink::AbstractParams: All values from sink will be assigned to corresponding fields of sink.

DEBBase.Utils.into — Method

into!(sink::AbstractParams, source::AbstractParams)

Put values from source into sink.

source::AbstractParams: A param struct from which we want to extract all parameters.

Sink::DataType: All values from sink will be assigned to a new instance of type Sink, assuming that a constructor Sink() exists.

DEBBase.Utils.ismin — Method

Get positions of minimum values as BitVector.

DEBBase.Utils.isolate_pmoas — Method

isolate_pmoas(spc::AbstractParams, pmoas::Vector{String}, z::Int64)::AbstractParams

Isolate the indicated PMoAs for chemical stressor z. That means, turn off all PMoAs (including lethal effects h) except for those indicated in pmoas-Vector. This is done through the toxicokinetic rate constant.

DEBBase.Utils.lab — Method

lab(v::Vector{R}; kwargs...) where R <: Real

Create legend labels from Vector of numeric Values. Kwargs are handed down to fround.

DEBBase.Utils.read_W3C — Method

read_W3C(file_path::AbstractString; kwargs...)::DataFrame

Read a csv including W3C-formatted metadata. Each line with metadata starts with a hashtag (#). For example:

#t: time in days
#W: wet weight in grams

t,W
0,0.1
1,0.2
2,0.3

Additional kwargs are handed down to CSV.File. The function currently returns just the core data.

DEBBase.Utils.relative_response — Method

Calculate the relative responses.

Positional arguments:

sim::AbstractDataFrame: results

response_vars::Vector{Symbol}: response variables for which to calculate the relative responses

treatment_var::Symbol: Column indicating the treatment. Column values can be numerical or categorical, but identify_control kwarg has to be specified in the latter case

Keyword arguments:

groupby_vars::Vector{Symbol}: relative response will be conditioned on these variables (e.g. time, separate experiments...). Empty by default.

identify_control: function used to identify reference values from treatment_var. By default, this is minimum() (assuming that column values in treatment_var are numerical).

DEBBase.Utils.replace_na! — Method

replace_na!(
    df::AbstractDataFrame,
    cols::Vector{Symbol};
    replace_val = 0.0
    )

Replace all missing values in DataFrame with replace_val.

DEBBase.Utils.rr — Method

Calculate relative response, given scalar values.

DEBBase.Utils.set_equal! — Method

Set parameters with common prefix to the value of a reference parameter. E.g.

set_equal!(spc, :Idot_max_rel, :lrv)

sets all parameters whose names start with Idot_max_rel equal to Idot_max_rel_lrv.

DEBBase.Utils.skipinf — Method

Skip inifite values.

DEBBase.Utils.sol_to_df — Method

Convert ODE solution object to output data frame.

DEBBase.Utils.sol_to_mat — Method

sol_to_mat(sol::O)::Matrix{Float64}

Convert ODE solution object to matrix.

DEBBase.Utils.trim! — Method

trim!(spc::AbstractParams)

Trim TKTD parameters to the smallest number of stressors indicated for any PMoA. For example k_D_G = [1., 0.], k_D_M = [1.] will be trimmed to k_D_G = [1.], k_D_M = [1.].

DEBBase.Utils.vectify — Method

Parse String to Vector of Floats.

DEBBase.Utils.which_in — Method

Identify element of possibilities occuring in x.
E.g., which_in("I like apples", ["apples", "bananas"]) returns "apples".
If multiple possibilities occur, first detection is returned.

DEBBase.Utils.wrappend — Method

wrappend(file::String, data::DataFrame, step::Int64)

Write DataFrames to disc during loop. Will overwrite existing file if step == 1 and append if step > 1.